+++ club-service/src/main/java/club/gach_dong/config/SwaggerConfig.java
@@ -3,6 +3,8 @@ package club.gach_dong.config;
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
 import io.swagger.v3.oas.models.servers.Server;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -15,9 +17,18 @@ public class SwaggerConfig {
     @Bean
     public OpenAPI openAPI() {
         return new OpenAPI()
-                .components(new Components())
                 .info(apiInfo())
-                .addServersItem(serverItem());
+                .addServersItem(serverItem())
+                .components(component());
+    }
+
+    private Components component() {
+        return new Components()
+                .addSecuritySchemes("Authorization",
+                        new SecurityScheme()
+                                .type(SecurityScheme.Type.HTTP)
+                                .scheme("Bearer")
+                                .bearerFormat("JWT"));
     }
 
     private Info apiInfo() {
