+++ club-service/src/main/java/club/gach_dong/controller/ClubController.java
@@ -9,8 +9,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 
-import java.util.List;
-
 @Controller
 @RequiredArgsConstructor
 public class ClubController implements ClubApiSpecification {
@@ -18,31 +16,8 @@ public class ClubController implements ClubApiSpecification {
     private final ClubService clubService;
 
     @Override
-    public ArrayResponse<ClubSummaryResponse> getClubs() {
-        List<ClubSummaryResponse> clubList = clubService.getAllClubs();
-        return ArrayResponse.of(clubList);
-    }
-
-    @Override
-    public ClubResponse getClub(String clubId) {
-        return clubService.getClub(clubId);
-    }
-
-    @Override
     public ResponseEntity<ClubResponse> createClub(CreateClubRequest createClubRequest) {
         ClubResponse clubResponse = clubService.createClub(createClubRequest);
         return new ResponseEntity<>(clubResponse, HttpStatus.CREATED);
-    }
-
-    @Override
-    public ArrayResponse<ClubActivityResponse> getClubActivities(String clubId) {
-        List<ClubActivityResponse> clubActivityResponse = clubService.getClubActivities(clubId);
-        return ArrayResponse.of(clubActivityResponse);
-    }
-
-    @Override
-    public ArrayResponse<ClubContactInfoResponse> getClubContactInfo(String clubId) {
-        List<ClubContactInfoResponse> clubContactInfoResponse = clubService.getClubContactInfo(clubId);
-        return ArrayResponse.of(clubContactInfoResponse);
     }
 }
