+++ club-service/src/main/java/club/gach_dong/controller/ClubPublicController.java
@@ -1,0 +1,42 @@
+package club.gach_dong.controller;
+
+import club.gach_dong.api.ClubPublicApiSpecification;
+import club.gach_dong.dto.response.ArrayResponse;
+import club.gach_dong.dto.response.ClubActivityResponse;
+import club.gach_dong.dto.response.ClubContactInfoResponse;
+import club.gach_dong.dto.response.ClubResponse;
+import club.gach_dong.dto.response.ClubSummaryResponse;
+import club.gach_dong.service.ClubService;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@RequiredArgsConstructor
+public class ClubPublicController implements ClubPublicApiSpecification {
+
+    private final ClubService clubService;
+
+    @Override
+    public ArrayResponse<ClubSummaryResponse> getClubs() {
+        List<ClubSummaryResponse> clubList = clubService.getAllClubs();
+        return ArrayResponse.of(clubList);
+    }
+
+    @Override
+    public ClubResponse getClub(String clubId) {
+        return clubService.getClub(clubId);
+    }
+
+    @Override
+    public ArrayResponse<ClubActivityResponse> getClubActivities(String clubId) {
+        List<ClubActivityResponse> clubActivityResponse = clubService.getClubActivities(clubId);
+        return ArrayResponse.of(clubActivityResponse);
+    }
+
+    @Override
+    public ArrayResponse<ClubContactInfoResponse> getClubContactInfo(String clubId) {
+        List<ClubContactInfoResponse> clubContactInfoResponse = clubService.getClubContactInfo(clubId);
+        return ArrayResponse.of(clubContactInfoResponse);
+    }
+}
